requirements.txt
pypdf
python-dotenv
chromadb
google-generativeai
sentence-transformers
langchain
umap-learn
matplotlib
pandas